flask-docker/requirements.txt
astroid==2.2.5
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pylint==2.3.1
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2
Flask
Flask-Restful
Flask-SQLAlchemy
psycopg2-binary
python-dotenv

gunicorn
flask_admin
WTForms
Flask-SuperAdmin